feat: migrate controller.js to controller.ts
